backend/requirements.txt
# StelleWorld Backend Requirements

# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Paiements Stripe
stripe==7.7.0

# HTTP et WebSocket
httpx==0.25.2
websockets==12.0

# Tâches asynchrones
celery==5.3.4
redis==5.0.1

# Notifications
python-telegram-bot==20.7

# Email
fastapi-mail==1.4.1

# Validation et configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilitaires
python-slugify==8.0.1
pillow==10.1.0
qrcode==7.4.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring et logs
structlog==23.2.0

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0
